XC-230: move test queue_holds_one_copy_of_each_task
